backend/requirements.txt
# Core dependencies
Flask==2.2.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
requests==2.28.2
pydantic==1.10.7

# API integrations
slack-sdk==3.20.2
notion-client==2.0.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.81.0

# Testing
pytest==7.3.1
pytest-mock==3.10.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3 